First pass refactoring scrapers into separate modules
